change toggleDropdown to reducer implemintation
