+++ src/widgets/ProfilePreferencesForm/ui/ProfilePreferencesForm.tsx
@@ -1,11 +1,15 @@
 import React, { FC } from "react";
 
-import styles from "./ProfilePreferencesForm.module.scss";
 import {
-    Activity, PopularPlaces, AcrossRussia, InputSelectedCountries,
+    AcrossRussia,
+    Activity,
+    InputSelectedCountries,
+    PopularPlaces,
 } from "@/features/ProfilePreferences";
 
-export const ProfilePreferencesForm:FC = () => (
+import styles from "./ProfilePreferencesForm.module.scss";
+
+export const ProfilePreferencesForm: FC = () => (
     <div className={styles.wrapper}>
         <PopularPlaces />
         <AcrossRussia />
